Removes the QFileIconProvider from TagFileSystemModel
The QFileIconProvider providing the icons for directory checking has an
inherited cache which cannot be emptied. Apparently, the
QFileSystemWatcher used by QFileSystemModel does not clean the cache
neither. So the icons are provided by the data() method which makes the
usage more flexible.

Additionally, the code is cleaned up a bit.